### 2019-08-29 #1690 @labarba
👋  @knaidoo29 —  I am the Associate Editor-in-Chief on rotation this week. I'm going to send this back to you for some more work, because the paper does not quite meet [our expectations](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). The paper, as submitted, is very brief and doesn't give enough of a description of the software or its need and uses in research. We don't expect a full article, but we recommend 250–1000 words (your Summary is just under that). These items need to be included:

- the summary describing high-level functionality and purpose of the software for a diverse, non-specialist audience;
- a clear statement of need that illustrates the research application of the software (including references to other software meeting similar needs).

TL;DR: the article needs to be improved with more description of how the software works, some examples, and use cases.
