### 2019-08-31 #1690 @knaidoo29
Hi @labarba, thank you for the comments. I am currently working on a significant revision/rewrite of the paper and was wondering (1) what the timescale for edits could be (I am planning to have this ready within a week) and (2) what the process was for resubmitting (would I just need to push my new version of the paper to the software's repository?).
