### 2019-08-31 #1690 @labarba
You can just push your changes to the paper, and let us know here when ready. I'll set a reminder for you.
