### 2019-09-05 331182f
adding binning and plotting classes
