+++ paper/paper.bib
@@ -1,3 +1,82 @@
+@article{spada2004,
+  title={Use of the minimum spanning tree model for molecular epidemiological investigation of a nosocomial outbreak of hepatitis C virus infection},
+  author={Spada, Enea and Sagliocca, Luciano and Sourdis, John and Garbuglia, Anna Rosa and Poggi, Vincenzo and De Fusco, Carmela and Mele, Alfonso},
+  journal={Journal of clinical microbiology},
+  volume={42},
+  number={9},
+  pages={4230--4236},
+  year={2004},
+  publisher={Am Soc Microbiol}
+}
+
+@articles{prim,
+  author={R. C. {Prim}},
+  journal={The Bell System Technical Journal},
+  title={Shortest connection networks and some generalizations},
+  year={1957},
+  volume={36},
+  number={6},
+  pages={1389-1401},
+  keywords={},
+  doi={10.1002/j.1538-7305.1957.tb01515.x},
+  ISSN={0005-8580},
+  month={Nov},}
+
+@article{rainbolt:2016,
+	Archiveprefix = {arXiv},
+	Author = {Rainbolt, Jessica Lovelace and Schmitt, Michael},
+	Date-Added = {2017-02-17 18:52:54 +0000},
+	Date-Modified = {2017-02-17 18:58:48 +0000},
+	Doi = {10.1088/1748-0221/12/02/P02009},
+	Eprint = {1608.04772},
+	Journal = {JINST},
+	Number = {02},
+	Pages = {P02009},
+	Primaryclass = {stat.AP},
+	Reportnumber = {NUHEP-EX-16-04},
+	Slaccitation = {%%CITATION = ARXIV:1608.04772;%%},
+	Title = {{The Use of Minimal Spanning Trees in Particle Physics}},
+	Volume = {12},
+	Year = {2017},
+	Bdsk-Url-1 = {http://dx.doi.org/10.1088/1748-0221/12/02/P02009}}
+
+@article{gama:2014,
+	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	Adsurl = {http://adsabs.harvard.edu/abs/2014MNRAS.438..177A},
+	Archiveprefix = {arXiv},
+	Author = {{Alpaslan}, M. and {Robotham}, A.~S.~G. and {Driver}, S. and {Norberg}, P. and {Baldry}, I. and {Bauer}, A.~E. and {Bland-Hawthorn}, J. and {Brown}, M. and {Cluver}, M. and {Colless}, M. and {Foster}, C. and {Hopkins}, A. and {Van Kampen}, E. and {Kelvin}, L. and {Lara-Lopez}, M.~A. and {Liske}, J. and {Lopez-Sanchez}, A.~R. and {Loveday}, J. and {McNaught-Roberts}, T. and {Merson}, A. and {Pimbblet}, K.},
+	Date-Added = {2017-02-17 19:00:00 +0000},
+	Date-Modified = {2017-02-17 19:00:13 +0000},
+	Doi = {10.1093/mnras/stt2136},
+	Eprint = {1311.1211},
+	Journal = {\mnras},
+	Keywords = {methods: observational, surveys, large-scale structure of Universe},
+	Month = feb,
+	Pages = {177-194},
+	Primaryclass = {astro-ph.CO},
+	Title = {{Galaxy And Mass Assembly (GAMA): the large-scale structure of galaxies and comparison to mock universes}},
+	Volume = 438,
+	Year = 2014,
+	Bdsk-Url-1 = {http://dx.doi.org/10.1093/mnras/stt2136}}
+
+@article{allison:2009,
[…]
