+++ paper/paper.md
@@ -19,12 +19,61 @@ bibliography: paper.bib
 
 # Summary
 
-``MiSTree`` is a ``Python`` package for the construction and analysis of *minimum spanning trees* (MST) from points given in 2/3 dimensions and in tomographic (on a unit sphere) or spherical polar coordinates. The weights of the edges are assumed to be the distances between points; i.e. the Euclidean distance for 2/3 dimension and spherical polar coordinates, and angular distances for tomographic coordinates. The package was designed to be used in cosmology; to extract cosmic web information present in the large scale distribution of galaxies (shown in @naidoo:2019) but could in principle be constructed on any distribution of points with non-Gaussian features.
+``MiSTree`` is a ``Python`` package for the fast construction of the *minimum spanning tree* (MST). ``MiSTree`` can take as input a distribution of points defined in different coordinate systems (2/3 dimensions, tomographic and spherical polar coordinates). A user can then use the constructed MST to look for patterns in the distribution or to measure and plot the MST statistics.
 
-The MST is constructed by initially creating a k-nearest neighbour graph (using ``scikit-learn``'s ``kneighbours_graph`` function) and then runs the Kruskal algorithm [@kruskal1956shortest] (using ``scipy`` ``minimum_spanning_tree`` function). Statistics of the constructed MST can then be measured and are described in @naidoo:2019.
+# Motivation
 
-The package was designed to be easy to use, with the class function ``GetMST`` allowing for the construction and subsequent analysis of the MST to be carried out fairly quickly and on a variety of data sets. A user can then choose to either measure the default statistics, discussed in @naidoo:2019, or develop their own.
+Studies of point distributions often measure their 2-point statistics (i.e. the distribution of distances between pairs of points) which are then compared to theoretical models. This is a powerful technique and has been used very successfully in the field of cosmology to study the early Universe and the large scale distribution of galaxies. Unfortunately this statistic can only fully describe a distribution that is Gaussian, if it is non-Gaussian then the 2-point is no longer sufficient. The conventional method to incorporate non-Gaussian information is to look at the distribution's $N$-point statistic (if $N\!=\!3$ we look at the distribution of triangles, if $N\!=\!4$ we look at the distribution of quadrilaterals and so on). This method is well motivated as in principle all the information that can describe a distribution of points is contained within its $N$-point statistics. However, calculating $N$-point statistics even for $N\!>\!3$ becomes quickly intractable for large data sets.
 
-Dependencies include the ``Python`` modules ``numpy`` [@numpy], ``scipy`` [@scipy], ``scikit-learn`` [@scikit-learn] and ``f2py`` [@peterson2009f2py] (the latter of which is used to compile ``Fortran`` subroutines). The source code can be found on [github](https://github.com/knaidoo29/mistree) and documentation and tutorials are provided [here](https://knaidoo29.github.io/mistreedoc/).
[…]
