+++ mistree/mst/hist_mst.py
@@ -1,0 +1,239 @@
+import numpy as np
+import matplotlib.pylab as plt
+
+
+def bin_data(data, minimum=None, maximum=None, bin_size=None, bin_number=100, normalised=True):
+    """Returns the (normalised) number count of a data set with values within defined bins.
+
+    Parameters
+    ----------
+    data : array_like
+        The data to be binned.
+    minimum : float
+        A given minimum for the bin edges.
+    maximum : float
+        A given maximum for the bin edges
+    bin_size : float
+        The size of the bins.
+    bin_number : int
+        The number of bins to be used.
+    normalised : bool
+        Tf true will return a normalised histogram, if false returns a number counts.
+
+    Returns
+    -------
+    bin_centres : array_like
+        The central value of each bin.
+    binned_data : array_like
+        The binned number count (or normalised histogram) of the data set.
+    """
+    if minimum is None:
+        minimum = np.min(data)
+    if maximum is None:
+        maximum = np.max(data)
+    if bin_size is None:
+        _bin_edge = np.linspace(minimum, maximum, bin_number+1)
+    binned_data, _bin_edges = np.histogram(data, bins=_bin_edge, density=normalised)
+    bin_centres = 0.5 * (_bin_edge[1:] + _bin_edge[:-1])
+    return bin_centres, binned_data
+
+
+class HistMST:
+
+    def __init__(self):
+        """Initialises MST histogram class.
+        """
+        # bin edges information
+        self.uselog = None
+        self.use_s_sqrt = None
+        self.usenorm = None
+        self.d_min = None
+        self.d_max = None
+        self.num_d_bins = None
+        self.l_min = None
+        self.l_max = None
+        self.num_l_bins = None
+        self.b_min = None
+        self.b_max = None
+        self.num_b_bins = None
+        self.s_min = None
+        self.s_max = None
+        self.num_s_bins = None
+        self.logl_min = None
+        self.logl_max = None
+        self.logb_min = None
+        self.logb_max = None
+        self.group_mode = False
+
+    def setup(self, uselog=False, use_sqrt_s=True, usenorm=True, d_min=0.5, d_max=6.5, num_d_bins=6,
+              l_min=0., l_max=None, num_l_bins=100, b_min=0., b_max=None, num_b_bins=100,
+              s_min=0., s_max=1., num_s_bins=50, logl_min=None, logl_max=None, logb_min=None, logb_max=None):
+        """Setups bin sizes for the MST statistics.
+
+        Parameters
+        ----------
+        uselog : bool
+            Determines whether to use log bins for l and b.
+        use_sqrt_s : bool
+            Determines whether to use the sqrt(1-s) projection of s or just s itself.
+        usenorm : bool
+            Determines whether to normalise the histograms.
+        d_min : float
+            Minimum for degree bins (use half integer values).
+        d_max : float
+            Maximum for degree bins (use half integer values).
+        num_d_bins : int
+            Number of bins for the distribution of degree, this should be equal to d_max - d_min.
+        l_min : float
+            Minimum for edge length bins.
+        l_max : float
+            Maximum for edge length bins.
+        num_l_bins : int
+            Number of bins for the distribution of edge lengths.
+        b_min : float
+            Minimum for branch length bins.
+        b_max : float
+            Maximum for branch length bins.
+        num_b_bins : int
+            Number of bins for the distribution of branch lengths.
[…]

[…]
