+++ paper/paper.md
@@ -25,13 +25,13 @@ bibliography: paper.bib
 
 Studies of point distributions often measure their 2-point statistics (i.e. the distribution of distances between pairs of points) which are then compared to theoretical models. This is a powerful technique and has been used very successfully in the field of cosmology to study the early Universe and the large scale distribution of galaxies. Unfortunately this statistic can only fully describe a distribution that is Gaussian, if it is non-Gaussian then the 2-point is no longer sufficient. The conventional method to incorporate non-Gaussian information is to look at the distribution's $N$-point statistic (if $N\!=\!3$ we look at the distribution of triangles, if $N\!=\!4$ we look at the distribution of quadrilaterals and so on). This method is well motivated as in principle all the information that can describe a distribution of points is contained within its $N$-point statistics. However, calculating $N$-point statistics even for $N\!>\!3$ becomes quickly intractable for large data sets.
 
-The MST offers an alternative approach; the MST graph draws lines between pairs of points so that all points are linked in a single skeletal structure that contains no loops and has minimal total edge length. Unlike $N$-point statistics, that typically scale by $\mathcal{O}(n^{N})$ for $n$ points, the MST (computed using the Kruskal algorithm @kruskal1956shortest) can be constructed much faster (at best $\mathcal{O}(n\log n)$). While the MST does not contain all the information present in $N$-point statistics, it enables some of this information to be captured and allows the identification of skeletal patterns, as such it has found a broad range of applications in physics: such as finding filaments in the distribution of galaxies [@gama:2014], classifying particle physics collisions [@rainbolt:2016] and mass segregation in star clusters [@allison:2009]. The MST has also been used in a number of other scientific field such as computer science, sociology and epidemiology.
+The MST offers an alternative approach; the MST graph draws lines between pairs of points so that all points are linked in a single skeletal structure that contains no loops and has minimal total edge length. Unlike $N$-point statistics, that typically scale by $\mathcal{O}(n^{N})$ for $n$ points, the MST (computed using the Kruskal algorithm [@kruskal1956shortest]) can be constructed much faster (at best $\mathcal{O}(n\log n)$). While the MST does not contain all the information present in $N$-point statistics, it enables some of this information to be captured and allows the identification of skeletal patterns, as such it has found a broad range of applications in physics: such as finding filaments in the distribution of galaxies [@gama:2014], classifying particle physics collisions [@rainbolt:2016] and mass segregation in star clusters [@allison:2009]. The MST has also been used in a number of other scientific field such as computer science, sociology and epidemiology.
 
 While algorithms to construct the *minimum spanning tree* are well known (e.g. @prim and @kruskal1956shortest) implementations of these often require the input of a matrix of pairwise distances. For a large data set the creation of this matrix (with $n^{2}$ elements) can be a significant strain on memory while also making the construction of the MST slower ($\mathcal{O}(n^{2}\log n)$).
 
 # MiSTree
 
-``MiSTree`` tries to solve these issues by providing a public ``Python`` package for the construction and analysis of the MST. The package initially creates a $k$-nearest neighbour graph ($k$NN, a graph that links each point to the nearest $k$ neighbours, using ``scikit-learn``'s ``kneighbours_graph`` function) which improves speed by limiting the number of considered edges from $n^{2}$ to $kn$ (where $k\!\l\!l n$) and then runs the Kruskal algorithm [@kruskal1956shortest] (using ``scipy``s' ``minimum_spanning_tree`` function). It was designed to be easy to use, with the raw input data into ``scipy``s' ``minimum_spanning_tree`` function and output being handled by ```MiSTree``` itself.
[…]
