### 2019-09-06 #1690 @knaidoo29
Hi @labarba and @xuanxu just to let you know I've updated the paper and pushed it to the repository. I rewrote it and have added some basic examples.

Note, that I've added a new version of the code (i.e. v1.1) which includes some new functions for binning and plotting the Minimum Spanning Tree statistics. I'm not familiar with how you will review the code but I just wanted to highlight this as I believe I placed v1.0 as the version to be tested.

I look forward to your comments.
