### 2019-09-06 #1690 @knaidoo29
Hi @xuanxu, I just had a look at the proofs and realised my figures are not there. This was because I had my gitignore set up to ignore png files. The images are now on the repository so if the pdf is regenerated it should appear on it now. Thanks (p.s can I regenerate the PDF or do I have to wait for you to do it?)
