### 2019-09-06 #1690 @labarba
You can also run `@whedon generate pdf` (as a single line in a separate comment) to compile the proof at any time.
