### 2019-09-12 knaidoo29/mistree#1 @zonca
Missing requirement
I created an environment to test `mistree` with `conda`:

    conda create -n test_mistree python=3.7 numpy scipy scikit-learn

I can run `setup.py` fine, but at import I get:

    ModuleNotFoundError: No module named 'matplotlib'

If I install `matplotlib`, the library imports fine.
please add `matplotlib` to the requirements or make its import optional.
