### 2019-09-12 knaidoo29/mistree#2 @zonca
Package is not installable
The python package is not installable. it is designed to be executed from its root folder. I recommend @knaidoo29 to make a installable Python package. I think the quickest way is to start from a template like the astropy package template, http://docs.astropy.org/projects/package-template/. This would also make it easy to setup integration with Travis-CI for testing.
