### 2019-09-12 #1721 @zonca
## Installation

* missing requirement on `matplotlib` https://github.com/knaidoo29/mistree/issues/1
* package is not installable https://github.com/knaidoo29/mistree/issues/2
