### 2019-09-12 be12770
added pip install instructions and setup
