### 2019-09-13 25ad3ca
removed old setup file
