+++ README.md
@@ -28,7 +28,9 @@ The `--user` is optional and only required if you don’t have write permission.
 want to work on the Github version you can clone the repository and install an editable version:
 
 `git clone https://github.com/knaidoo29/mistree.git`
+
 `cd mistree`
+
 `pip install -e . [--user]`
 
 You should now be able to import the module:
