+++ README.md
@@ -22,7 +22,9 @@ You will need the following python modules:
 
 MiSTree can be installed as follows:
 
-`pip install mistree [--user]`
+```
+pip install mistree [--user]
+```
 
 The `--user` is optional and only required if you don’t have write permission. If you
 want to work on the Github version you can clone the repository and install an editable version::
