+++ README.md
@@ -12,6 +12,7 @@ Python 3.5 and 3.7.
 
 You will need the following python modules:
 
+* Python 2.7 or 3.5+
 * `numpy`
 * `matplotlib`
 * `scipy`
