### 2019-09-17 c228358
add test scripts
