+++ mistree/tests/test_branches.py
@@ -1,0 +1,162 @@
+import numpy as np
+import mistree as mist
+
+
+def test_get_branch_index():
+    x = np.random.random_sample(50)
+    y = np.random.random_sample(50)
+    mst = mist.GetMST(x=x, y=y)
+    mst.construct_mst()
+    mst.get_degree()
+    mst.get_degree_for_edges()
+    branch_index, branch_index_rejected = mist.get_branch_index(mst.edge_index, mst.edge_degree)
+    assert len(branch_index_rejected) == 0
+
+
+def test_get_branch_index_sub_divide_2d():
+    x = np.random.random_sample(50)
+    y = np.random.random_sample(50)
+    mst = mist.GetMST(x=x, y=y)
+    mst.construct_mst()
+    mst.get_degree()
+    mst.get_degree_for_edges()
+    branch_index, branch_index_rej = mist.get_branch_index(mst.edge_index, mst.edge_degree)
+    branch_index_sub, branch_index_sub_rej = mist.get_branch_index_sub_divide(2, mst.edge_index, mst.edge_degree,
+                                                                              box_size=1., edge_x=mst.edge_x, edge_y=mst.edge_y)
+    assert len(branch_index_sub_rej) == 0
+    count = 0
+    for i in range(0, len(branch_index)):
+        for j in range(0, len(branch_index_sub)):
+            if np.array_equal(np.array(sorted(branch_index[i])), np.array(sorted(branch_index_sub[j]))) == True:
+                count += 1
+            else:
+                pass
+    assert count == len(branch_index)
+
+
+def test_get_branch_index_sub_divide_3d():
+    x = np.random.random_sample(50)
+    y = np.random.random_sample(50)
+    z = np.random.random_sample(50)
+    mst = mist.GetMST(x=x, y=y, z=z)
+    mst.construct_mst()
+    mst.get_degree()
+    mst.get_degree_for_edges()
+    branch_index, branch_index_rej = mist.get_branch_index(mst.edge_index, mst.edge_degree)
+    branch_index_sub, branch_index_sub_rej = mist.get_branch_index_sub_divide(2, mst.edge_index, mst.edge_degree,
+                                                                              box_size=1., edge_x=mst.edge_x, edge_y=mst.edge_y, edge_z=mst.edge_z,
+                                                                              two_dimension=False)
+    assert len(branch_index_sub_rej) == 0
+    count = 0
+    for i in range(0, len(branch_index)):
+        for j in range(0, len(branch_index_sub)):
+            if np.array_equal(np.array(sorted(branch_index[i])), np.array(sorted(branch_index_sub[j]))) == True:
+                count += 1
+            else:
+                pass
+    assert count == len(branch_index)
+
+
+def test_get_branch_index_sub_divide_3d_no_box_size():
+    x = np.random.random_sample(50)
+    y = np.random.random_sample(50)
+    z = np.random.random_sample(50)
+    mst = mist.GetMST(x=x, y=y, z=z)
+    mst.construct_mst()
+    mst.get_degree()
+    mst.get_degree_for_edges()
+    branch_index, branch_index_rej = mist.get_branch_index(mst.edge_index, mst.edge_degree)
+    branch_index_sub, branch_index_sub_rej = mist.get_branch_index_sub_divide(2, mst.edge_index, mst.edge_degree,
+                                                                              edge_x=mst.edge_x, edge_y=mst.edge_y, edge_z=mst.edge_z)
+    assert len(branch_index_sub_rej) == 0
+    count = 0
+    for i in range(0, len(branch_index)):
+        for j in range(0, len(branch_index_sub)):
[…]
