+++ .travis.yml
@@ -1,0 +1,13 @@
+language: python
+python:
+  - "2.7"
+  - "3.4"
+  - "3.5"
+  - "3.6"
+
+install:
+  - pip install -r numpy matplotlib scipy scikit-learn
+
+script:
+  - nose
+ 
