+++ .travis.yml
@@ -6,7 +6,7 @@ python:
   - "3.6"
 
 install:
-  - pip install -r numpy matplotlib scipy scikit-learn
+  - pip install -r requirements.txt
 
 script:
   - nose
