+++ .travis.yml
@@ -1,4 +1,6 @@
 language: python
+cache: pip
+
 python:
   - "2.7"
   - "3.4"
