+++ .travis.yml
@@ -9,6 +9,7 @@ python:
  
 install:
   - pip install -r requirements.txt
+  - pip install .
 
 script: nosetests
  
