+++ .travis.yml
@@ -6,7 +6,10 @@ python:
   - "3.4"
   - "3.5"
   - "3.6"
- 
+
+before_install:
+  - apt-get install gfortran
+
 install:
   - pip install -r requirements.txt
   - pip install .
