+++ .travis.yml
@@ -8,7 +8,7 @@ python:
   - "3.6"
 
 before_install:
-  - apt-get install gfortran
+  - sudo apt-get install gfortran
 
 install:
   - pip install -r requirements.txt
