+++ .travis.yml
@@ -14,5 +14,9 @@ install:
   - pip install -r requirements.txt
   - pip install .
 
+before_script:
+  - ls
+  - cd mistree
+
 script: nosetests
  
