+++ .travis.yml
@@ -12,6 +12,7 @@ before_install:
 
 install:
   - pip install -r requirements.txt
+  - pip install codecov
   - pip install .
 
 before_script:
@@ -22,4 +23,4 @@ before_script:
 script: python setup.py test
 
 after_success:
-  - coveralls
+  - codecov
