+++ .travis.yml
@@ -23,5 +23,5 @@ before_script:
 script: python setup.py test
 
 after_success:
-  - coverage run --source=mypkg setup.py test
+  - coverage run --source=mistree setup.py test
   - codecov
