+++ mistree/tests/test_plot_mst.py
@@ -1,3 +1,5 @@
+import matplotlib
+matplotlib.use('Agg')
 import numpy as np
 import mistree as mist
 
