+++ mistree/tests/test_plot_mst.py
@@ -1,5 +1,5 @@
 import matplotlib
-matplotlib.use('Agg')
+matplotlib.use('TkAgg')
 import numpy as np
 import mistree as mist
 
