+++ .travis.yml
@@ -16,10 +16,10 @@ install:
   - pip install .
 
 before_script:
-  - ls /home/travis/build/knaidoo29/mistree/mistree/levy_flight
-  - ls /home/travis/build/knaidoo29/mistree/mistree/mst
-  - ls /home/travis/build/knaidoo29/mistree/mistree/coordinates
-
+  - "export DISPLAY=:99.0"
+  - "sh -e /etc/init.d/xvfb start"
+  - sleep 3 # give xvfb some time to start
+  
 script: python setup.py test
 
 after_success:
