+++ mistree/tests/test_construct.py
@@ -1,4 +1,6 @@
 import numpy as np
+import matplotlib
+matplotlib.use('agg')
 import mistree as mist
 
 
