+++ .travis.yml
@@ -15,11 +15,6 @@ install:
   - pip install codecov
   - pip install .
 
-before_script:
-  - "export DISPLAY=:99.0"
-  - "sh -e /etc/init.d/xvfb start"
-  - sleep 3 # give xvfb some time to start
-  
 script: python setup.py test
 
 after_success:
