+++ mistree/tests/test_levy_flight.py
@@ -1,6 +1,4 @@
 import numpy as np
-import matplotlib
-matplotlib.use('agg')
 import mistree as mist
 
 
