+++ .travis.yml
@@ -1,5 +1,9 @@
 language: python
+
 cache: pip
+
+services:
+  - xvfb
 
 python:
   - "2.7"
