+++ README.md
@@ -3,7 +3,7 @@
 Author:         Krishna Naidoo                          
 Version:        1.1                                     
 Homepage:       https://github.com/knaidoo29/mistree    
-Documentation:  https://knaidoo29.github.io/mistreedoc/ 
+Documentation:  https://knaidoo29.github.io/mistreedoc/
 
 [![Build Status](https://travis-ci.org/knaidoo29/mistree.svg?branch=master)](https://travis-ci.org/knaidoo29/mistree) [![codecov](https://codecov.io/gh/knaidoo29/mistree/branch/master/graph/badge.svg)](https://codecov.io/gh/knaidoo29/mistree) [![PyPI version](https://badge.fury.io/py/mistree.svg)](https://badge.fury.io/py/mistree) [![status](https://joss.theoj.org/papers/461d79e9e5faf21029c0a7b1c928be28/status.svg)](https://joss.theoj.org/papers/461d79e9e5faf21029c0a7b1c928be28) [![DOI](https://zenodo.org/badge/170473458.svg)](https://zenodo.org/badge/latestdoi/170473458) [![License: MIT](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/licenses/MIT)
 
@@ -12,7 +12,7 @@ analysing minimum spanning trees.
 
 ## Dependencies
 
-* Python 2.7 or 3.5+
+* Python 2.7 or 3.4+
 * `numpy`
 * `matplotlib`
 * `scipy`
