### 2019-09-17 knaidoo29/mistree#7 @zonca
Improve JOSS paper
- [x] Expand summary, add clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience
- [x] After you fix #4 , see if you can re-use some of that in the paper
- [x] You mention `n-point` statistics, can you look for a review paper on that you can cite?

I'll expand this in the next days but you can start working on these items.
