+++ README.md
@@ -29,7 +29,7 @@ pip install mistree [--user]
 ```
 
 The `--user` is optional and only required if you don’t have write permission. If you
-want to work on the Github version you can clone the repository and install an editable version::
+want to work on the Github version you can clone the repository and install an editable version:
 
 ```
 git clone https://github.com/knaidoo29/mistree.git
@@ -50,4 +50,4 @@ In depth documentation and tutorials are provided [here](https://knaidoo29.githu
 ## Support
 
 If you have any issues with the code or want to suggest ways to improve it please open a new issue (here: https://github.com/knaidoo29/mistree/issues)
-or (if you don't have a github account) email _krishna.naidoo.11@ucl.ac.uk_
+or (if you don't have a github account) email _krishna.naidoo.11@ucl.ac.uk_.
