### 2019-09-18 8e5ad80
add ipython notebook tutorial
