### 2019-09-18 #1721 @zonca
ok, @knaidoo29, anything related to software looks fine to me, now we'll focus on the paper: https://github.com/knaidoo29/mistree/issues/7
