### 2019-09-18 aa53b41
add ipython notebook for mst
