### 2019-09-19 5bec2b3
add environment for binder
