+++ environment.yml
@@ -1,0 +1,5 @@
+dependencies:
+  - numpy
+  - matplotlib
+  - scipy
+  - scikit-learn
