### 2019-09-19 6a6f674
update environment for binder
