+++ environment.yml
@@ -1,0 +1,7 @@
+name: mistree
+
+dependencies:
+  - numpy
+  - scipy
+  - matplotlib
+  - scikit-learn
