+++ environment.yml
@@ -5,3 +5,5 @@ dependencies:
   - scipy
   - matplotlib
   - scikit-learn
+  - pip:
+    - mistree
