+++ paper/paper.md
@@ -19,11 +19,11 @@ bibliography: paper.bib
 
 # Summary
 
-``MiSTree`` is a ``Python`` package for the fast construction of the *minimum spanning tree* (MST). ``MiSTree`` can take as input a distribution of points defined in different coordinate systems (2/3 dimensions, tomographic and spherical polar coordinates). A user can then use the constructed MST to look for patterns in the distribution or to measure and plot the MST statistics.
+The *minimum spanning tree* (MST) has been used in a broad range of scientific research including computer science, epidemiology, social sciences, particle physics, astronomy and cosmology. Its success in these field has been driven by its sensitivity to the spatial distribution of points and the patterns within. ``MiSTree``, a public ``Python`` package, allows a user to construct the MST in a variety of coordinates systems, including Celestial coordinates used in astronomy. The package enables the MST to be constructed quickly by initially using a k-nearest neighbour graph (rather than a matrix of pairwise distances) which is then fed to Kruskal's algorithm to construct the MST. ``MiSTree`` enables a user to measure the statistics of the MST and provides classes for binning the MST statistics (into histograms) and plotting the distributions. Including the MST in cosmological parameter estimation studies will enable the inclusion of high-order statistics information from the cosmic web. This information has traditionally been unexploited due to the computational cost of calculating N-point statistics.
 
 # Motivation
 
-Studies of point distributions often measure their 2-point statistics (i.e. the distribution of distances between pairs of points) which are then compared to theoretical models. This is a powerful technique and has been used very successfully in the field of cosmology to study the early Universe and the large scale distribution of galaxies. Unfortunately this statistic can only fully describe a distribution that is Gaussian, if it is non-Gaussian then the 2-point is no longer sufficient. The conventional method to incorporate non-Gaussian information is to look at the distribution's $N$-point statistic (if $N\!=\!3$ we look at the distribution of triangles, if $N\!=\!4$ we look at the distribution of quadrilaterals and so on). This method is well motivated as in principle all the information that can describe a distribution of points is contained within its $N$-point statistics. However, calculating $N$-point statistics even for $N\!>\!3$ becomes quickly intractable for large data sets.
+Studies of point distributions often measure their 2-point statistics (i.e. the distribution of distances between pairs of points) which are then compared to theoretical models. This is a powerful technique and has been used very successfully in the field of cosmology to study the early Universe and the large scale distribution of galaxies. Unfortunately this statistic can only fully describe a distribution that is Gaussian, if it is non-Gaussian then the 2-point is no longer sufficient. The conventional method to incorporate non-Gaussian information is to look at the distribution's $N$-point statistic (if $N\!=\!3$ we look at the distribution of triangles, if $N\!=\!4$ we look at the distribution of quadrilaterals and so on). This method is well motivated as in principle all the information that can describe a distribution of points is contained within its $N$-point statistics (see @szapudi1998). However, calculating $N$-point statistics even for $N\!>\!3$ becomes quickly intractable for large data sets.
 
 The MST offers an alternative approach; the MST graph draws lines between pairs of points so that all points are linked in a single skeletal structure that contains no loops and has minimal total edge length. Unlike $N$-point statistics, that typically scale by $\mathcal{O}(n^{N})$ for $n$ points, the MST (computed using the Kruskal algorithm [@kruskal1956shortest]) can be constructed much faster (at best $\mathcal{O}(n\log n)$). While the MST does not contain all the information present in $N$-point statistics, it enables some of this information to be captured and allows the identification of skeletal patterns, as such it has found a broad range of applications in physics: such as finding filaments in the distribution of galaxies [@gama:2014], classifying particle physics collisions [@rainbolt:2016] and mass segregation in star clusters [@allison:2009]. The MST has also been used in a number of other scientific field such as computer science, sociology and epidemiology.
 
