+++ paper/paper.bib
@@ -1,3 +1,18 @@
+@ARTICLE{szapudi1998,
+       author = {{Szapudi}, Istv{\'a}n and {Szalay}, Alexander S.},
+        title = "{A New Class of Estimators for the N-Point Correlations}",
+      journal = {\apjl},
+     keywords = {COSMOLOGY: LARGE-SCALE STRUCTURE OF UNIVERSE, METHODS: STATISTICAL, GALAXIES: STATISTICS, COSMOLOGY: THEORY, Cosmology: Theory, Galaxies: Statistics, Cosmology: Large-Scale Structure of Universe, Methods: Statistical},
+         year = "1998",
+        month = "Feb",
+       volume = {494},
+       number = {1},
+        pages = {L41-L44},
+          doi = {10.1086/311146},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1998ApJ...494L..41S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @article{spada2004,
   title={Use of the minimum spanning tree model for molecular epidemiological investigation of a nosocomial outbreak of hepatitis C virus infection},
   author={Spada, Enea and Sagliocca, Luciano and Sourdis, John and Garbuglia, Anna Rosa and Poggi, Vincenzo and De Fusco, Carmela and Mele, Alfonso},
