### 2019-09-20 knaidoo29/mistree#7 @knaidoo29
Hi @zonca I've replaced the summary with the introduction I provide in the README and documentation. I've also added a reference to an implementation of n-point statistics by Szapudi and Szalay (1998).
