### 2019-09-25 knaidoo29/mistree#7 @zonca
- [ ] in the abstract you are mentioning Kruskal’s algorithm without any link or explanation.
- [ ] you say: "...will enable the inclusion of high-order statistics information from the cosmic web", go ahread, what does this high order statistics allow you to do that wasn't doable before?
