### 2019-10-01 knaidoo29/mistree#7 @melissawm
I agree with the points above and will wait for this issue to close before I approve the paper. Otherwise it looks good to me.
