### 2019-10-02 knaidoo29/mistree#7 @knaidoo29
Hi both, apologies about the delay. I've made some changes to the paper following @zonca suggestions.

I now provide a reference for the kruskal algorithm and provide a brief explanation in the second paragraph in the motivation section.

I mention in the abstract that adding higher order statistics allows for improved constraints on cosmological parameters.

As for the target audience I've added this line at the end of the summary:

> MiSTree was designed to be used in cosmology but could be used in any field which requires extracting non-Gaussian information from point distributions.

And I've changed the typo in the spelling of facilities.
