### 2019-10-02 knaidoo29/mistree#7 @knaidoo29
Oh, I forgot to mention I've made some cosmetic changes to the plots included to make them look a bit better and to showcase the capabilities of the plotting routines.
