### 2019-10-05 knaidoo29/mistree#7 @zonca
thanks @knaidoo29, I think we are almost done, please check a few other points below:

- [ ] improve summary: the paper starts with "The minimum spanning tree (MST) has been used in a broad range of scientific fields including computer science, epidemiology, social sciences, particle physics, astronomy and cosmology", please first explain what is a minimum spanning tree is and to what kind of data is applied. Then instead of a list of disciplines, I think it would be better to just mention 2 or 3 and for each of them a simple example of an application so the reader can get an idea of the kind of problems a MST can solve
- [ ] typo:  `scipys'` -> `scipy's`
- [ ] please rephrase this, I don't understand the meaning: "It was designed to be easy to use, with the raw input data into scipys' minimum_spanning_tree function and output being handled by MiSTree itself"
- [ ] you mention this has been used in cosmology, please cite a paper about that
