+++ paper/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-The *minimum spanning tree* (MST) has been used in a broad range of scientific fields including computer science, epidemiology, social sciences, particle physics, astronomy and cosmology. Its success in these field has been driven by its sensitivity to the spatial distribution of points and the patterns within. ``MiSTree``, a public ``Python`` package, allows a user to construct the MST in a variety of coordinates systems, including Celestial coordinates used in astronomy. The package enables the MST to be constructed quickly by initially using a $k$-nearest neighbour graph ($k$NN, rather than a matrix of pairwise distances) which is then fed to Kruskal's algorithm [@kruskal1956shortest] to construct the MST. ``MiSTree`` enables a user to measure the statistics of the MST and provides classes for binning the MST statistics (into histograms) and plotting the distributions. Including the MST will enable the inclusion of high-order statistics information from the cosmic web which can provide additional information to improve cosmological parameter constraints. This information has traditionally been unexploited due to the computational cost of calculating $N$-point statistics. ``MiSTree`` was designed to be used in cosmology but could be used in any field which requires extracting non-Gaussian information from point distributions.
+The *minimum spanning tree* (MST) is a graph constructed from a distribution of points. The MST draws lines between pairs of points so that all points are linked in a single skeletal structure that contains no loops and has minimal total edge length. The MST has been used in a broad range of scientific fields such as particle physics (to distinguish classes of events in collider collisions, see @rainbolt:2016), in astronomy (to detect mass segregation in star clusters) and cosmology (to search for filaments in the cosmic web, see @gama:2014). Its success in these fields has been driven by its sensitivity to the spatial distribution of points and the patterns within. ``MiSTree``, a public ``Python`` package, allows a user to construct the MST in a variety of coordinates systems, including Celestial coordinates used in astronomy. The package enables the MST to be constructed quickly by initially using a $k$-nearest neighbour graph ($k$NN, rather than a matrix of pairwise distances) which is then fed to Kruskal's algorithm [@kruskal1956shortest] to construct the MST. ``MiSTree`` enables a user to measure the statistics of the MST and provides classes for binning the MST statistics (into histograms) and plotting the distributions. Including the MST will enable the inclusion of high-order statistics information from the cosmic web which can provide additional information to improve cosmological parameter constraints [@naidoo:2019]. This information has traditionally been unexploited due to the computational cost of calculating $N$-point statistics. ``MiSTree`` was designed to be used in cosmology but could be used in any field which requires extracting non-Gaussian information from point distributions.
 
 # Motivation
 
@@ -33,7 +33,7 @@ While algorithms to construct the *minimum spanning tree* are well known (e.g. @
 
 # MiSTree
 
-``MiSTree`` is a public ``Python`` package for the construction and analysis of the MST. The package initially creates a $k$-nearest neighbour graph ($k$NN, a graph that links each point to the nearest $k$ neighbours, using ``scikit-learn``'s ``kneighbours_graph`` function) which improves speed by limiting the number of considered edges from $n^{2}$ to $kn$ (where $k\!\ll\! n$) and then runs the Kruskal algorithm [@kruskal1956shortest] (using ``scipy``s' ``minimum_spanning_tree`` function). The stages of the MST construction are shown in Figure 1. It was designed to be easy to use, with the raw input data into ``scipy``s' ``minimum_spanning_tree`` function and output being handled by ```MiSTree``` itself.
[…]
