### 2019-10-11 f437b2c
fix branch shape bug
