+++ README.md
@@ -1,7 +1,7 @@
 # MiSTree
 
 Author:         Krishna Naidoo                          
-Version:        1.1.3                               
+Version:        1.1.4                               
 Homepage:       https://github.com/knaidoo29/mistree    
 Documentation:  https://knaidoo29.github.io/mistreedoc/
 
