### 2019-10-15 #1721 @knaidoo29
Great to see that it's been accepted! Really appreciate all the help from both @zonca and @melissawm. I think the code and the paper are now much stronger.

@xuanxu is there anything I need to do before publication?
