### 2019-10-16 #1721 @xuanxu
@knaidoo29: Can you take a look at the list of DOIs missing info and add them if [the suggestions](#issuecomment-542829879) from whedon are corect?
