+++ paper/paper.bib
@@ -13,15 +13,20 @@
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{spada2004,
-  title={Use of the minimum spanning tree model for molecular epidemiological investigation of a nosocomial outbreak of hepatitis C virus infection},
-  author={Spada, Enea and Sagliocca, Luciano and Sourdis, John and Garbuglia, Anna Rosa and Poggi, Vincenzo and De Fusco, Carmela and Mele, Alfonso},
-  journal={Journal of clinical microbiology},
-  volume={42},
-  number={9},
-  pages={4230--4236},
-  year={2004},
-  publisher={Am Soc Microbiol}
+@article {spada2004,
+	author = {Spada, Enea and Sagliocca, Luciano and Sourdis, John and Garbuglia, Anna Rosa and Poggi, Vincenzo and De Fusco, Carmela and Mele, Alfonso},
+	title = {Use of the Minimum Spanning Tree Model for Molecular Epidemiological Investigation of a Nosocomial Outbreak of Hepatitis C Virus Infection},
+	volume = {42},
+	number = {9},
+	pages = {4230--4236},
+	year = {2004},
+	doi = {10.1128/JCM.42.9.4230-4236.2004},
+	publisher = {American Society for Microbiology Journals},
+	abstract = {The minimum spanning tree (MST) model was applied to identify the history of transmission of hepatitis C virus (HCV) infection in an outbreak involving five children attending a pediatric oncology-hematology outpatient ward between 1992 and 2000. We collected blood samples from all children attending since 1992, all household contacts, and one health care worker positive for antibody to HCV (anti-HCV). HCV RNA detection was performed with these samples and with smears of routinely collected bone marrow samples. For all isolates, we performed sequence analysis and phylogenetic tree analysis of hypervariable region 1 of the E2 gene. The MST model was applied to clinical-epidemiological and molecular data. No additional cases were detected. All children, but not the health care worker, showed genotype 3a. On six occasions, all but one child had shared the medication room with another patient who later seroconverted. HCV RNA detection in bone marrow smears revealed, in some cases, a delay of several months in anti-HCV responses. Sequence analysis and phylogenetic tree analysis revealed a high identity among the isolates. The MST model applied to molecular data, together with the clinical-epidemiological data, allowed us to identify the source of the outbreak and the most probable patient-to-patient chain of transmission. The management of central venous catheters was suspected to be the probable route of transmission. In conclusion, the MST model, supported by an exhaustive clinical-epidemiological investigation, appears to be a useful tool in tracing the history of transmission in outbreaks of HCV infection.},
+	issn = {0095-1137},
+	URL = {https://jcm.asm.org/content/42/9/4230},
+	eprint = {https://jcm.asm.org/content/42/9/4230.full.pdf},
+	journal = {Journal of Clinical Microbiology}
 }
 
 @articles{prim,
@@ -156,14 +161,22 @@ archivePrefix = {arXiv},
 }
 
 @article{peterson2009f2py,
-  title={F2PY: a tool for connecting Fortran and Python programs},
-  author={Peterson, Pearu},
-  journal={International Journal of Computational Science and Engineering},
-  volume={4},
-  number={4},
-  pages={296--305},
-  year={2009},
-  publisher={Inderscience Publishers}
+ author = {Peterson, Pearu},
+ title = {F2PY; a Tool for Connecting Fortran and Python Programs},
+ journal = {Int. J. Comput. Sci. Eng.},
[…]
