+++ mistree/mst/plot_mst.py
@@ -264,7 +264,7 @@ class PlotHistMST:
         self.need_envelopes.append(mst_hist['isgroup'])
 
     def plot(self, usebox=True, saveas=None, fontsize=16, figsize=(16, 4), subplot_setup='4x1',
-             units=None, showenvelopes=True, usecomp=False, usemean=True, height_ratios=[2, 1],
+             units=None, showenvelopes=True, showsigma=2, usecomp=False, usemean=True, height_ratios=[2, 1],
              usefraction=False, whichcomp=0, plotzeroline=True, legend=True, subplot_adjust_top=0.85,
              legend_fontsize=14, legend_column=4, xlabels=[None, None, None, None], dpi=None, plt_output='show'):
         """Outputs the final plot of the MST statistics.
@@ -286,6 +286,8 @@ class PlotHistMST:
             Units of l and b MST statistics, if None is supplied then we assume it is unitless.
         showenvelopes : bool
             This determines whether to plot data with input standard deviation.
+        showsigma : int
+            Number of sigma errorbars to show.
         usecomp : bool
             Determines whether to include comparison subplots.
         usemean : bool
@@ -443,8 +445,9 @@ class PlotHistMST:
                                 ax=ax1, color=self.colors[i], alpha=self.alphas[i],
                                 linewidth=self.linewidths[i], linestyle=self.linestyles[i])
             if showenvelopes == True and self.need_envelopes[i] == True:
-                plot_histogram_error(self.binned_data[i]['x_d'], self.binned_data[i]['y_d'], self.binned_data[i]['y_d_std'],
-                                     ax=ax1, color=self.colors[i], alpha=self.alphas_envelope[i])
+                for sigma in range(0, showsigma):
+                    plot_histogram_error(self.binned_data[i]['x_d'], self.binned_data[i]['y_d'], (1+sigma)*self.binned_data[i]['y_d_std'],
+                                         ax=ax1, color=self.colors[i], alpha=self.alphas_envelope[i]/(1+sigma))
             if usebox == True:
                 plot_histogram_line(self.binned_data[i]['x_l'], self.binned_data[i]['y_l'], x_edges=self.l_edges,
                                     ax=ax2, color=self.colors[i], alpha=self.alphas[i],
@@ -456,12 +459,13 @@ class PlotHistMST:
                                     ax=ax4, color=self.colors[i], alpha=self.alphas[i],
                                     linewidth=self.linewidths[i], linestyle=self.linestyles[i])
                 if showenvelopes == True and self.need_envelopes[i] == True:
-                    plot_histogram_error(self.binned_data[i]['x_l'], self.binned_data[i]['y_l'], self.binned_data[i]['y_l_std'],
-                                         x_edges=self.l_edges, ax=ax2, color=self.colors[i], alpha=self.alphas_envelope[i])
-                    plot_histogram_error(self.binned_data[i]['x_b'], self.binned_data[i]['y_b'], self.binned_data[i]['y_b_std'],
-                                         x_edges=self.b_edges, ax=ax3, color=self.colors[i], alpha=self.alphas_envelope[i])
-                    plot_histogram_error(self.binned_data[i]['x_s'], self.binned_data[i]['y_s'], self.binned_data[i]['y_s_std'],
-                                         ax=ax4, color=self.colors[i], alpha=self.alphas_envelope[i])
+                    for sigma in range(0, showsigma):
+                        plot_histogram_error(self.binned_data[i]['x_l'], self.binned_data[i]['y_l'], (1+sigma)*self.binned_data[i]['y_l_std'],
+                                             x_edges=self.l_edges, ax=ax2, color=self.colors[i], alpha=self.alphas_envelope[i]/(1+sigma))
[…]
