### 2019-10-16 #1721 @knaidoo29
Hi @xuanxu, just checked these doi's and they are correct. I've uploaded them and added them to the bib file.
