### 2019-10-16 #1721 @xuanxu
OK @knaidoo29, everything looks good, here are the next steps:
- Please release a new version from the current master so it includes all the changes made during the review process
- Then archive that latest release in Zenodo
- Check the Zenodo deposit has the correct metadata: title and author name should match the paper; you may also add your ORCID.

Once you do that please report here the Zenodo DOI.
