### 2019-10-17 #1721 @knaidoo29
Okay @xuanxu I've created a new version (v1.2) which is archived on zenodo with the doi: 10.5281/zenodo.3495008 and has matching title to the paper, author and orcid.
