### 2019-10-17 #1721 @xuanxu
Everything looks ready 🎉
Pinging @openjournals/joss-eics for final acceptance
