### 2019-10-17 #1721 @Kevin-Mattheus-Moerman
@openjournals/jose-eics does anybody know why for [this paper](https://github.com/openjournals/joss-papers/blob/joss.01721/joss.01721/10.21105.joss.01721.pdf) the "matplotlib" project name is broken over 2 lines without a hyphen? is this because it is seen as a piece of code text? Since they use the `` `` `` symbols? i.e. the authors use
```mardown
Dependencies for ``MiSTree`` include the ``Python`` modules ``numpy`` [@numpy], ``matplotlib`` [@Hunter:2007]
```
In the paper this is broken onto the next line like this:
> Dependencies for MiSTree include the Python modules ``numpy`` (Oliphant, 2006), ``matplo``
> ``tlib`` (Hunter, 2007), ...

The authors have used code style formatting for all project names. If this is the issue then perhaps a simple quick fix is to use _italic_ fonts rather then `code` fonts for project names.
