### 2019-10-17 #1721 @kyleniemeyer
@Kevin-Mattheus-Moerman yes, that's probably what's going on. I wonder if adding a line break (`\\`) before the word would help?
