+++ paper/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-The *minimum spanning tree* (MST), a graph constructed from a distribution of points, draws lines between pairs of points so that all points are linked in a single skeletal structure that contains no loops and has minimal total edge length. The MST has been used in a broad range of scientific fields such as particle physics (to distinguish classes of events in collider collisions, see @rainbolt:2016), in astronomy (to detect mass segregation in star clusters, see @allison:2009) and cosmology (to search for filaments in the cosmic web, see @gama:2014). Its success in these fields has been driven by its sensitivity to the spatial distribution of points and the patterns within. ``MiSTree``, a public ``Python`` package, allows a user to construct the MST in a variety of coordinates systems, including Celestial coordinates used in astronomy. The package enables the MST to be constructed quickly by initially using a $k$-nearest neighbour graph ($k$NN, rather than a matrix of pairwise distances) which is then fed to Kruskal's algorithm [@kruskal1956shortest] to construct the MST. ``MiSTree`` enables a user to measure the statistics of the MST and provides classes for binning the MST statistics (into histograms) and plotting the distributions. Applying the MST will enable the inclusion of high-order statistics information from the cosmic web which can provide additional information to improve cosmological parameter constraints [@naidoo:2019]. This information has not been fully exploited due to the computational cost of calculating $N$-point statistics. ``MiSTree`` was designed to be used in cosmology but could be used in any field which requires extracting non-Gaussian information from point distributions.
+The *minimum spanning tree* (MST), a graph constructed from a distribution of points, draws lines between pairs of points so that all points are linked in a single skeletal structure that contains no loops and has minimal total edge length. The MST has been used in a broad range of scientific fields such as particle physics (to distinguish classes of events in collider collisions, see @rainbolt:2016), in astronomy (to detect mass segregation in star clusters, see @allison:2009) and cosmology (to search for filaments in the cosmic web, see @gama:2014). Its success in these fields has been driven by its sensitivity to the spatial distribution of points and the patterns within. *MiSTree*, a public *Python* package, allows a user to construct the MST in a variety of coordinates systems, including Celestial coordinates used in astronomy. The package enables the MST to be constructed quickly by initially using a $k$-nearest neighbour graph ($k$NN, rather than a matrix of pairwise distances) which is then fed to Kruskal's algorithm [@kruskal1956shortest] to construct the MST. *MiSTree* enables a user to measure the statistics of the MST and provides classes for binning the MST statistics (into histograms) and plotting the distributions. Applying the MST will enable the inclusion of high-order statistics information from the cosmic web which can provide additional information to improve cosmological parameter constraints [@naidoo:2019]. This information has not been fully exploited due to the computational cost of calculating $N$-point statistics. *MiSTree* was designed to be used in cosmology but could be used in any field which requires extracting non-Gaussian information from point distributions.
 
 # Motivation
 
@@ -29,11 +29,11 @@ The MST offers an alternative approach; the MST graph draws lines between pairs 
 
 While algorithms to construct the *minimum spanning tree* are well known (e.g. @prim and @kruskal1956shortest) implementations of these often require the input of a matrix of pairwise distances. For a large data set the creation of this matrix (with $n^{2}$ elements) can be a significant strain on memory while also making the construction of the MST slower ($\mathcal{O}(n^{2}\log n)$).
 
[…]
