### 2019-10-17 #1721 @knaidoo29
Hi @Kevin-Mattheus-Moerman just noticed that the footnote starts with `Naidoo, (3030).` shouldn't this be `Naidoo, (2019).` or `Naidoo, K. (2019).`?
